Use Alfred's "Open URL" action for pin and annotate
Big Sur's `open -g <url>` doesn't do any URL encoding.
The one on Catalina does. As a consequence, the following
works on Catalina, but not on Big Sur:
  open -g "cleanshot://share-extension-pin?filepath=/path/with spaces.png"

Given that the only reason to not use Alfred's bulit-in
"Open URL" action in the first place was because it has
weird behavior with "show/hide desktop icons", it makes sense
to just use that action for the pin/annotate cases.